feat: block delete patient

+++ backend/src/4 - Infrastructure/persistence/prisma/repositories/PatientRepository.ts
@@ -1,48 +1,26 @@
 import { Patient } from "../../../../3 - Domain/entities/Patient";
-import { IPatientRepository} from "../../../../3 - Domain/repositories/IPatientRepository";
+import { IPatientRepository } from "../../../../3 - Domain/repositories/IPatientRepository";
 
 import prisma from "../PrismaClient";
 
 export class PatientRepository implements IPatientRepository {
-    async create(patient: Patient): Promise<Patient> {
-        const newPatient = await prisma.paciente.create({
-            data: {
-                id_paciente: patient.patientId,
-                nome_paciente: patient.name,
-                cpf_paciente: patient.cpfPatient,
-                telefone: patient.phone,
-                endereco: patient.address,
-                genero: patient.gender,
-                data_nasc: patient.birthday,
-                email: patient.email
-            },
-        })
-        .then(
-            (patient) =>
-                new Patient(
-                    patient.id_paciente,
-                    patient.nome_paciente,
-                    patient.cpf_paciente,
-                    patient.telefone,
-                    patient.endereco,
-                    patient.genero,
-                    patient.data_nasc,
-                    patient.email
-                )
-        );
-        return newPatient;
-    }
-
-    async findById(patientId: number): Promise<Patient | null> {
-        const patient = await prisma.paciente.findUnique({
-            where: { id_paciente: patientId }, // 
-        });
-
-        if (!patient) {
-            throw new Error("Paciente não foi encontrado!");
-        }
-
-        return new Patient(
+  async create(patient: Patient): Promise<Patient> {
+    const newPatient = await prisma.paciente
+      .create({
+        data: {
+          id_paciente: patient.patientId,
+          nome_paciente: patient.name,
+          cpf_paciente: patient.cpfPatient,
+          telefone: patient.phone,
+          endereco: patient.address,
+          genero: patient.gender,
+          data_nasc: patient.birthday,
+          email: patient.email,
+        },
+      })
+      .then(
+        (patient) =>
+          new Patient(
             patient.id_paciente,
             patient.nome_paciente,
             patient.cpf_paciente,
@@ -51,13 +29,67 @@ export class PatientRepository implements IPatientRepository {
             patient.genero,
             patient.data_nasc,
             patient.email
-        );
+          )
+      );
+    return newPatient;
+  }
+
+  async findById(patientId: number): Promise<Patient | null> {
+    const patient = await prisma.paciente.findUnique({
+      where: { id_paciente: patientId }, //
+    });
+
+    if (!patient) {
+      throw new Error("Paciente não foi encontrado!");
     }
 
-    async findAll(): Promise<Patient[]> {
-        const patients = await prisma.paciente.findMany();
+    return new Patient(
+      patient.id_paciente,
+      patient.nome_paciente,
+      patient.cpf_paciente,
+      patient.telefone,
+      patient.endereco,
+      patient.genero,
+      patient.data_nasc,
+      patient.email
+    );
+  }
 
-        return patients.map(patient => new Patient(
+  async findAll(): Promise<Patient[]> {
+    const patients = await prisma.paciente.findMany();
+
+    return patients.map(
+      (patient) =>
+        new Patient(
+          patient.id_paciente,
+          patient.nome_paciente,
+          patient.cpf_paciente,
+          patient.telefone,
+          patient.endereco,
+          patient.genero,
+          patient.data_nasc,
+          patient.email
+        )
+    );
+  }
+
+  async update(patient: Patient): Promise<Patient> {
+    const updatedPatient = await prisma.paciente
+      .update({
+        where: { id_paciente: patient.patientId },
+        data: {
+          nome_paciente: patient.name,
+          cpf_paciente: patient.cpfPatient,
+          telefone: patient.phone,
+          endereco: patient.address,
+          genero: patient.gender,
+          data_nasc: patient.birthday,
+          email: patient.email,
+        },
+      })
+      .then(
+        (patient) =>
+          new Patient(
             patient.id_paciente,
             patient.nome_paciente,
             patient.cpf_paciente,
@@ -66,42 +98,24 @@ export class PatientRepository implements IPatientRepository {
             patient.genero,
             patient.data_nasc,
             patient.email
-        ));
-    }
+          )
+      );
+    return updatedPatient;
+  }
 
-    async update(patient: Patient): Promise<Patient> {
-        const updatedPatient = await prisma.paciente
-            .update({
-                where: { id_paciente: patient.patientId },
-                data: {
-                    nome_paciente: patient.name,
-                    cpf_paciente: patient.cpfPatient,
-                    telefone: patient.phone,
-                    endereco: patient.address,
-                    genero: patient.gender,
-                    data_nasc: patient.birthday,
-                    email: patient.email
-                },
-            })
-            .then(
-                (patient) =>
-                    new Patient(
-                        patient.id_paciente,
-                        patient.nome_paciente,
-                        patient.cpf_paciente,
-                        patient.telefone,
-                        patient.endereco,
-                        patient.genero,
-                        patient.data_nasc,
-                        patient.email
-                    )
-            );
-        return updatedPatient;
-    }
+  async delete(id: number): Promise<void> {
+    const hasScheduling = await prisma.agendamento.findFirst({
+      where: {
+        paciente_id_paciente: id,
+      },
+    });
 
-    async delete(id: number): Promise<void> {
-        await prisma.paciente.delete({
-            where: { id_paciente: id },
-        });
+    if (hasScheduling) {
+      throw new Error("Não é possível excluir paciente pois possui vínculos.");
     }
+    
+    await prisma.paciente.delete({
+      where: { id_paciente: id },
+    });
+  }
 }
